+++ bark_ml/library_wrappers/lib_tf_agents/runners/__init__.py
@@ -1,2 +1,3 @@
 from bark_ml.library_wrappers.lib_tf_agents.runners.sac_runner import SACRunner
+from bark_ml.library_wrappers.lib_tf_agents.runners.sac_runner import SACRunnerGenerator
 from bark_ml.library_wrappers.lib_tf_agents.runners.ppo_runner import PPORunner
